back/requirements.txt
aiokafka==0.12.0
fastapi==0.115.11
PyJWT==2.10.1
uvicorn==0.34.0
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
passlib==1.7.4
bcrypt==4.1.2
